Revert "Feature Add Cli Payment Processor"
This reverts commit f9d9733327e6667f98e824a739e6424d18721ad9.

+++ src/main/java/PaymentSystem/Payment.java
@@ -1,15 +1,14 @@
 package PaymentSystem;
 
-import java.math.BigDecimal;
 import java.util.*;
 
 public class Payment {
     private final String id;
-    private BigDecimal amount;
+    private int amount;
     private String currency;
     private PaymentStatus status;
 
-    public Payment(BigDecimal amount, String currency) {
+    public Payment(int amount, String currency) {
         this.id = UUID.randomUUID().toString();
         this.amount = amount;
         this.currency = currency;
@@ -19,7 +18,7 @@ public class Payment {
     public String getId() {
         return id;
     }
-    public BigDecimal getAmount() {
+    public int getAmount() {
         return amount;
     }
     public String getCurrency() {
